pyproject.toml
[project]
name = "backend"
version = "0.2.0"
description = "GraphGuard SaaS — Fraud Detection Backend"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "aiokafka>=0.12.0",
    "mesa>=3.5.1",
    "neo4j>=6.1.0",
    "networkx>=3.6.1",
    "pandas>=3.0.2",
    "python-dotenv>=1.2.2",
    "scikit-learn>=1.8.0",
    "torch>=2.11.0",
    "torch-geometric>=2.7.0",
]
